[DM-23776] Move form code into the token handler
Forms are intrinsically part of the web UI, so move the form code
into the tokens handler to slowly improve the layers of separation
in this code.
